+++ lib/src/event/dom.dart
@@ -529,7 +529,7 @@ class Dom {
    * @param elem the element whose listener is to be set
    * @return the element's event listener
    */
-  static dart_html.EventListener getEventListener(dart_html.Element elem) {
+  static EventListener getEventListener(dart_html.Element elem) {
     return domHelper.getEventListener(elem);
   }
 }