+++ lib/src/event/dom_helper.dart
@@ -125,5 +125,5 @@ abstract class DomHelper {
   
   dart_html.Event createHtmlEvent(String type, bool canBubble, bool cancelable);
   
-  dart_html.EventListener getEventListener(dart_html.Element elem);
+  EventListener getEventListener(dart_html.Element elem);
 }