+++ lib/src/event/impl/dom_helper_default.dart
@@ -184,7 +184,7 @@ class DomHelperDefault extends DomHelper {
 //    dart_html.window.on.gesturechange.add(dispatchCapturedMouseEvent, true);
   }
   
-  dart_html.EventListener getEventListener(dart_html.Element elem) {
+  EventListener getEventListener(dart_html.Element elem) {
     //return elem.__listener;
     return _listener.containsKey(elem) ? _listener[elem] : null;
   }